validación alumno y materia

+++ src/controllers/index.controller.js
@@ -18,13 +18,24 @@ const getAlumno = async (req, res) => {
 
 const getAlumnoById = async (req, res) => {
     const id = req.params.id;
+    const {name, secondName, groupId, promedio, age, status} = req.body;
+    console.log('id', id);
+    if (id === undefined||name === undefined||secondName === undefined||groupId === undefined||promedio === undefined||age === undefined||status === undefined){
+        return res.json({
+            message: 'Valores Faltantes'
+        })            
+    } 
     const response = await pool.query('SELECT * FROM alumno WHERE id = $1', [id]);
     res.json(response.rows);
 }
 
 const createAlumno = async (req, res) => {
     const { name, secondName, groupId, promedio, age, status } = req.body;
-
+    if (name === undefined||secondName === undefined||groupId === undefined||promedio === undefined||age === undefined||status === undefined){
+        return res.json({
+            message: 'Valores Faltantes'
+        })            
+    } 
     const response = await pool.query('INSERT INTO getAlumno(name, secondName, groupId, promedio, age, status) VALUES ($1, $2, $3, $4, $5, $6)', [name, secondName, groupId, promedio, age, status]);
     console.log(response);
     res.json({
@@ -38,6 +49,12 @@ const createAlumno = async (req, res) => {
 const updateAlumno = async (req, res) => {
     const id = req.params.id;
     const { name, secondName, groupId, promedio, age, status } = req.body;
+    console.log('id', id);
+    if (id === undefined||name === undefined||secondName === undefined||groupId === undefined||promedio === undefined||age === undefined||status === undefined){
+        return res.json({
+            message: 'Valores Faltantes'
+        })            
+    } 
     const response = await pool.query('UPDATE alumno SET name = $1, secondName = $2 WHERE id = $3', [
         name,
         secondName,
@@ -49,6 +66,13 @@ const updateAlumno = async (req, res) => {
 
 const deleteAlumno = async (req, res) => {
     const id = req.params.id;
+    const {name, secondName, groupId, promedio, age, status} = req.body;
+    console.log('id', id);
+    if (id === undefined||name === undefined||secondName === undefined||groupId === undefined||promedio === undefined||age === undefined||status === undefined){
+        return res.json({
+            message: 'Valores Faltantes'
+        })            
+    } 
     const response = await pool.query('DELETE FROM alumno WHERE id = $1', [id]);
     console.log(response);
     res.json(`Alumno ${id} deleted successfully`);
@@ -63,16 +87,27 @@ const getMateria = async (req, res) => {
 
 const getMateriaById = async (req, res) => {
     const id = req.params.id;
+    const {idAlumno, name, score} = req.body;
+    console.log('id', id);
+    if (id === undefined||idAlumno === undefined||name === undefined||score === undefined){
+        return res.json({
+            message: 'Valores Faltantes'
+        })            
+    } 
     const response = await pool.query('SELECT * FROM materia WHERE id = $1', [id]);
     res.json(response.rows);
 }
 
 const createMateria = async (req, res) => {
     const { idAlumno, name, score } = req.body;
-
+    if (idAlumno === undefined||name === undefined||score === undefined){
+        return res.json({
+            message: 'Valores Faltantes'
+        })            
+    } 
     const response = await pool.query('INSERT INTO materia (idAlumno, name, score) VALUES ($1, $2, $3)', [idAlumno, name, score]);
     console.log(response);
-    res.json({
+    return res.json({
         message: 'Materia Added Succesfully',
         body: {
             materia: {idAlumno, name, score}
@@ -83,6 +118,11 @@ const createMateria = async (req, res) => {
 const updateMateria = async (req, res) => {
     const id = req.params.id;
     const { idAlumno, name, score } = req.body;
+    if (idAlumno === undefined||name === undefined||score === undefined){
+        return res.json({
+            message: 'Valores Faltantes'
+        })            
+    } 
     const response = await pool.query('UPDATE materia SET idAlumno = $1, name = $2, score = $3 WHERE id = $4', [
         idAlumno,
         name,
@@ -95,6 +135,13 @@ const updateMateria = async (req, res) => {
 
 const deleteMateria = async (req, res) => {
     const id = req.params.id;
+    const {idAlumno, name, score} = req.body;
+    console.log('id', id);
+    if (id === undefined||idAlumno === undefined||name === undefined||score === undefined){
+        return res.json({
+            message: 'Valores Faltantes'
+        })            
+    } 
     const response = await pool.query('DELETE FROM materia WHERE id = $1', [id])
     console.log(response);
     res.json(`Materia ${id} deleted succesfully`);
